PT. Transjakarta Data Analysis focusing on Elderly Passenger
The success of this analysis will be measured by the ability to provide actionable recommendations that can be implemented to enhance the transportation experience for elderly/senior passengers. Ultimately, the aim is to position Transjakarta as a preferred mode of transport for the elderly, thereby promoting their mobility, independence, and quality of life.